+++ employee-first-project/src/main/java/com/increff/employee/dto/ProductDto.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.increff.employee.model.BrandCategoryData;
+import com.increff.employee.model.InventoryData;
 import com.increff.employee.pojo.BrandCategoryPojo;
 import com.increff.employee.pojo.InventoryPojo;
 import com.increff.employee.service.InventoryService;
@@ -25,19 +26,31 @@ public class ProductDto {
 	@Autowired
 	private InventoryService inventoryService;
 	@Autowired
+	private  InventoryDto inventoryDto;
+	@Autowired
 	private BrandCategoryDto brandCategoryDto;
 
 	public void add(ProductForm form) throws ApiException {
 		normalize(form);
 		BrandCategoryPojo brandCategoryPojo = brandCategoryDto.get(form.getBrand(),form.getCategory());
 		ProductPojo p = convert(form,brandCategoryPojo);
+		ProductPojo pojo = service.get(p.getBarcode());
+		if(pojo != null)
+		{
+			throw  new ApiException("This barcode is already used try some different");
+		}
 		int id=service.add(p);
 		InventoryPojo p2 = new InventoryPojo();
 		p2.setId(id);
 		inventoryService.add(p2);
 	}
 
-	public void delete(int id) {
+	public void delete(int id) throws ApiException {
+		InventoryData d = inventoryDto.get(id);
+		if(d.getInventory()>0)
+		{
+			throw new ApiException("You can't delete this product its inventory is not empty");
+		}
 		service.delete(id);
 	}
 
@@ -53,13 +66,15 @@ public class ProductDto {
 	public List<ProductPojo> getByBrandCategoryID(int BrandCategoryId){
 		  return service.getByBrandCategoryID(BrandCategoryId);
 	}
-	public ProductPojo get(String barcode) throws ApiException {
+
+	public ProductData get(String barcode) throws ApiException {
 	    ProductPojo p = service.get(barcode);
 		if(p==null)
 		{
 			throw new ApiException("Product not exist in inventory");
 		}
-		return p;
+		BrandCategoryData brandCategoryData = brandCategoryDto.get(p.getBrandCategoryId());
+		return convert(p,brandCategoryData);
 	}
 
 	public List<ProductData> getAll() throws ApiException {
